requirements.txt
flask==3.0.2
transformers==4.38.2
torch==2.2.1
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
gunicorn==21.2.0
python-dotenv==1.0.1
python-magic==0.4.27
pytest==8.0.2
